Updated example with login callback
