+++ login-workflow/example/src/App.tsx
@@ -65,6 +65,7 @@ export const App = (): JSX.Element => {
             <AppContext.Provider
                 value={{
                     isAuthenticated,
+                    setIsAuthenticated,
                     onUserAuthenticated: (userData): void => {
                         setIsAuthenticated(true);
                         setLoginData(userData);
